# Optimize performance of encaseP