--- a/src/encase-p3.js
+++ b/src/encase-p3.js
@@ -1,5 +1,5 @@
 import {Core} from './core';
-import {noop, show, showf, escapeTick, partial1, partial2, partial3} from './internal/fn';
+import {noop, show, showf, immediately, partial1, partial2, partial3} from './internal/fn';
 import {isThenable, isFunction} from './internal/is';
 import {invalidArgument, typeError} from './internal/throw';
 
@@ -24,7 +24,7 @@ EncaseP3.prototype = Object.create(Core);
 
 EncaseP3.prototype._fork = function EncaseP3$fork(rej, res){
   const {_fn, _a, _b, _c} = this;
-  check$promise(_fn(_a, _b, _c), _fn, _a, _b, _c).then(escapeTick(res), escapeTick(rej));
+  check$promise(_fn(_a, _b, _c), _fn, _a, _b, _c).then(immediately(res), immediately(rej));
   return noop;
 };