--- a/src/internal/bc.js
+++ b/src/internal/bc.js
@@ -1,0 +1,1 @@
+export const setImmediate = (f, x) => setTimeout(f, 0, x);