**Before**

```text
$ node bench/encase-p.js 
encaseP x 766 ops/sec ±0.68% (57 runs sampled)
encaseP2 x 778 ops/sec ±0.89% (79 runs sampled)
encaseP3 x 775 ops/sec ±0.71% (79 runs sampled)
```

(with an average CPU load of about 2%)

**After**

```text
$ node bench/encase-p.js 
encaseP x 398,000 ops/sec ±0.44% (83 runs sampled)
encaseP2 x 357,165 ops/sec ±4.89% (76 runs sampled)
encaseP3 x 364,562 ops/sec ±1.56% (82 runs sampled)
```

(with an average CPU load of about 80%)

---

Resolves #148